// src/shared/dataprovider/api/model/clientConfigureFeesDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VendorConfigureFeesDTO } from './vendorConfigureFeesDTO';
import { WithdrawalFeeConfigureDTO } from './withdrawalFeeConfigureDTO';

export interface ClientConfigureFeesDTO { 
    Id?: number;
    Name?: string;
    Vendors?: Array<VendorConfigureFeesDTO>;
    WithdrawalFees?: Array<WithdrawalFeeConfigureDTO>;
}