// src/shared/dataprovider/api/model/merchantConfigureFeesDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ConfigureFeeDTO } from './configureFeeDTO';
import { PaymentChannelConfigureFeeDTO } from './paymentChannelConfigureFeeDTO';
import { StatusEnum } from './statusEnum';

export interface MerchantConfigureFeesDTO { 
    Id?: number;
    Name?: string;
    Status?: StatusEnum;
    readonly VendorId?: number;
    CashInFee?: ConfigureFeeDTO;
    CashOutFee?: ConfigureFeeDTO;
    PaymentChannel?: PaymentChannelConfigureFeeDTO;
}